+++ charts/src/main/java/org/patternfly/chart/donut/DonutElement.java
@@ -20,12 +20,10 @@ import org.patternfly.chart.Data;
 
 import elemental2.core.JsArray;
 import jsinterop.annotations.JsPackage;
-import jsinterop.annotations.JsProperty;
 import jsinterop.annotations.JsType;
 
 @JsType(isNative = true, namespace = JsPackage.GLOBAL, name = "HTMLElement")
 public class DonutElement extends ChartElement {
 
-    @JsProperty
     public JsArray<Data> data;
 }