+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Bullet chart (#109)
+
 ## [0.4.1] - 2025-12-15
 
 ### Fixed